pre 4.23

+++ part4/bloglist/controllers/blogs.js
@@ -1,6 +1,7 @@
 const blogsRouter = require('express').Router()
 const Blog = require('../models/blog')
 const User = require('../models/user')
+const middleware = require("../utils/middleware")
 
 const jwt = require('jsonwebtoken')
 
@@ -9,23 +10,14 @@ blogsRouter.get('/', async (request, response) => {
     response.json(blogs)
 })
 
-
-const getTokenFrom = request => {
-    const authorization = request.get('authorization')
-    if (authorization && authorization.startsWith('Bearer ')) {
-        return authorization.replace('Bearer ', '')
-    }
-    return null
-}
-
 blogsRouter.post('/', async (request, response) => {
     const body = request.body
 
-    const decodedToken = jwt.verify(getTokenFrom(request), process.env.SECRET)
+    const decodedToken = jwt.verify(request.token, process.env.SECRET)
     if (!decodedToken.id) {
         return response.status(401).json({ error: 'token invalid' })
     }
-    const user = await User.findById(decodedToken.id)
+    const user = request.user
 
     const blog = new Blog({
         title: body.title,
@@ -36,6 +28,7 @@ blogsRouter.post('/', async (request, response) => {
     })
 
     const savedBlog = await blog.save()
+
     user.blogs = user.blogs.concat(savedBlog.id)
     await user.save()
 
@@ -43,6 +36,23 @@ blogsRouter.post('/', async (request, response) => {
 })
 
 blogsRouter.delete('/:id', async (request, response) => {
+
+    const user = request.user
+    const decodedToken = jwt.verify(request.token, process.env.SECRET)
+    if (!decodedToken.id) {
+        return response.status(401).json({ error: 'token invalid' })
+    }
+
+    const id = request.params.id
+    const blog = await Blog.findById(id)
+    console.log(blog.user.toString());
+    console.log(user.id.toString());
+    if (blog.user.toString() === user.id.toString()) {
+        await Blog.deleteOne({ id })
+        response.sendStatus(204).end()
+    } else {
+        response.status(401).json({ error: "unauthorized operation" })
+    }
     await Blog.findByIdAndRemove(request.params.id)
     response.status(204).end()
 })
